Add Current Timestamp as default for Created and Modified
Previous query provided a table with NULL defaults on `articles.created` and `articles.modified`. The tutorials then go on to create views displaying the date (index, view, and edit, etc.) using `format()`, which returns an error in 3.2+ for trying `format(NULL)`. On add and edit actions (from the tutorial), no date/time is passed with the title and body of the articles, leaving the value NULL in the db.

Simple fix. Probably piss off a lot less newcomers to Cake/PHP/MySQL.